requirements-gpu.txt
jupyter==1.0.0
notebook==6.0.2
numpy==1.17.4
matplotlib==3.1.1
tensorflow-gpu==1.15
tqdm==4.39.0
plotly==1.13
plotly_express==0.4.0
pandas==0.25
